Change Vector to use BigDecimal and add TestVector
Change Vector to store vector values in BigDecimal instead of doubles.
Alter various method parameters and return types throughout vector.
These major changes to the public API are typically bad, but since this
is before version 1.0.0, I've reserved the right to make major changes.
Alter Matrix and CartesianCoordinates accordingly to accommodate these
changes.

Move the test classes and resources to into package specific
directories. Add a section to the build.gradle to tell javadoc to link
to the java APIs build javadocs for this project.

Implement TestVector class and modified Vector to be more robust.

Add an overloaded method readFromDimacsFile that takes an InputStream
to make it easier to use project resources as input files for jUnit
tests.