Change removeNodeFromGraph to removeNode
This name is much better. I would simplify it even more, but it is
possible to remove both nodes and edges, so "Node" stays. Updated all
dependent files to reflect the change.

Add new test to TestUndirectedGraph to test getting complement in
boundary case where a node has no edges in the complement.