Add Graph::shortestPath
Implemented Dijkstra's algorithm using a priority queue. Modify
method Graph::getNodes() of the graph to return a collection that is
independent of the graph so that changes to it won't affect the graph.
Did the same thing in Node::getNeighbors(). Add unit tests to verify
this behavior. Add precondition check in Graph::getNode() to throw an
exception if the Node is not in the Graph. This necessitated changes
in IncMaxCliqueAdapter and IncMaxCliqueSolver. Add Graph::equals() and
Graph::hashCode().