Change linearAlgebra classes hierarchy
Streamline the hierarchy. Vector is a matrix, and Point3D is a Vector.
Add feature to Vector. They can be transposed. By default, they start as
column vectors. They can be transposed to become row vectors. This is
an important feature for matrix multiplication.

Rename CartesianCoordinates to Point3D, a much better name.