Add Transformations Feature
Add classes to allow 3D matrix transformations. A TransformationMatrix
is a [4 by 4] augmented matrix that can be used to transform Vectors in
3D space. It is an abstract class because any subclass needs to
implement inverse and the implementation depends on the type of
Transformation.

From a client's perspective, they care about Transformation and
Transformable. A Transformation is a TransformationMatrix and can
be composed out of Translations and Rotations. Both of these extend
TransformationMatrix and are private inner classes of Transformation.
For simplicity and encapsulation, they are hidden from clients. A
client instantiates a Transformation object and then adds rotations and
translations to it via the methods in Transformation. Transformations
can return their inverse. A Transformations inverse is the composition
of the inverses of all the Rotations and Translations that make it up.

Interface Transformable allows objects to be acted upon by
Transformations. Any class that implements Transformable must implement
the applyTransformation(Transformation t) method to allow itself to be
transformed. Point3D in this library is one such class.

Include JUnit class TestTransformation, tests each type of
transformation, including composed transformations and inverses.