Change Paramter type in Transformable
Change parameter type in Transformable::applyTransformation from
TransformationMatrix to Transformation. Transformation provides nice
features like keeping track of the history of Transformations that
have been added to it so that the inverse can nicely be calculated
at any time. Any object that should transform another should be
a Transformation. For future extensibility, new subclasses of
TransformationMatrix can be created, but Objects of type Transformation
should be the ones that are used to Transform Transformables.
Any instance of a subclass of TransformationMatrix can be added
to a Transformation and that Transformation can be applied to a
Transformable.