Change Numerical Representation in linearAlgebra
Change the underlying numerical representation from BigDecimals to
Doubles. The goal with BigDecimals was to increase numerical precision,
but running even simple matrix multiplication for Transformations took
far too long. For efficiency, the numbers were changed from BigDecimals
to Doubles.

BigDecimals are only used for equality checks because they
give a convenient way to set the precision and handle rounding. A double
is only precise to 15 digits, including those to the left and right of
the decimal. To check if two doubles are equals, determine how many digits
to the left of the decimal are reliable (from the perspective of double
representation). The last reliable digit is used for rounding. The
number of digits to the right of the decimal before that digit is
14 - # digits to left of decimal. Build a BigDecimal with that value as
the scale and use that big decimal for comparison.