Restrict Access to Graph Constructor taking Nodes
This library is moving to minimizing the direct access clients have to
Nodes. Nodes are an implementation detail. From a clients perspective,
a Graph consists of the elements they add to it (and Nodes are just
wrappers for them) and the connections between those elements.

Clients should not be building graphs from collections of Nodes. Rather
they should be building them from files or by adding edges or elements
to the graph. In practice the constructors that were restricted were
being used to build subsets of graphs. That should instead be done with
the subset methods, given a list of elements, subset returns a graph
containing those elements and the edges between them. This subset is
a deep copy of the original part of the graph.