Restrict Access to addNode in Graph
There is now only one and it accepts an element. The one that takes
a Node is private.

Clients should no longer be building graphs directly from the Node
objects themselves, whether through adding edges or directly adding the
nodes. That ability is restricted so that all side effects are handled
within the Graph class itself.