Decrease Unnecessary Reference Creation
In the MaxClique algorithms, it became clear that too much time and
space were being wasted on unnecessary references. For example, for
safety, in many places a copy of a graph, minus or including a set of
vertices, is created to be passed into a recursive call. Where possible,
instead of doing that, add or remove any vertices before the call. Then
undo the changes after. This eliminates an exponential number of Graph
instances being created in favor of one that is constantly changing.

Another example of expensive references is Node::getNeighbors(). For
safety, a new ArrayList was being instantiated and returned because
otherwise, changes to the returned Collection would alter the state of
the Node and of the Graph. Unfortunately, this method is called often.
Looping over a vertices neighbors is a very common operation. Much
less common is the need to modify that collection. As an efficiency
vs. safety trade-off, getNeighbors() now returns a reference to the
underlying Collection of neighbors but is made protected. It is not to
be used by clients, much as the Node class itself is generally not to be
directly accessed by clients. This one change itself roughly halved the
runtimes of the MausMaxClique and IncMaxClique algorithms over several
common test cases.

The required changes for usage was propagated through the classes.
Finally, a style change was made everywhere noticed, ") {" instead of
"){".